# Conflicts:
#	src/main/java/com/example/SWExhibition/api/CommentApiController.java
#	src/main/java/com/example/SWExhibition/controller/MoviesController.java

+++ src/main/java/com/example/SWExhibition/dto/RatingsDto.java
@@ -7,5 +7,5 @@ public class RatingsDto {
 
     private String userId;  // 유저 ID
     private String movieId; // 영화 ID
-    private Long rating;    // 평점
+    private Float rating;    // 평점
 }
